refactor(ddd): migrate core services into bounded context modules
