+++ koduck-backend/src/main/java/com/koduck/market/application/MarketServiceImpl.java
@@ -1,4 +1,4 @@
-package com.koduck.service.impl;
+package com.koduck.market.application;
 
 import com.koduck.config.CacheConfig;
 import com.koduck.dto.market.MarketIndexDto;
