+++ koduck-backend/src/main/java/com/koduck/market/application/TechnicalIndicatorServiceImpl.java
@@ -1,4 +1,4 @@
-package com.koduck.service.impl;
+package com.koduck.market.application;
 import com.koduck.common.constants.MarketConstants;
 import com.koduck.dto.indicator.IndicatorListResponse;
 import com.koduck.dto.indicator.IndicatorResponse;
