+++ koduck-backend/src/main/java/com/koduck/strategy/application/StrategyServiceImpl.java
@@ -1,4 +1,4 @@
-package com.koduck.service.impl;
+package com.koduck.strategy.application;
 
 import com.koduck.common.constants.MarketConstants;
 import com.koduck.dto.strategy.CreateStrategyRequest;
